requirement.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncmy
aiofiles
email-validator
python-dotenv
passlib[bcrypt]
python-jose
